add data_assets to computation api object
